FINAL vo2/FINAL vo2/requirements.txt
streamlit>=1.28.0
tensorflow>=2.15.0
numpy>=1.24.0
pillow>=10.0.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
matplotlib>=3.7.0
seaborn>=0.12.0

# OpenAI for LLM chat integration
openai>=1.0.0
